+++ packages/transport/src/config.ts
@@ -1,6 +1,0 @@
-export const DEFAULT_URL = 'http://127.0.0.1:21325';
-// todo: DEFAULT_VERSION_URL is not really used anywhere in production at the moment
-export const DEFAULT_VERSION_URL = 'https://data.trezor.io/bridge/latest.txt';
-export const MESSAGE_HEADER_BYTE = 0x23;
-export const HEADER_SIZE = 1 + 1 + 4 + 2;
-export const BUFFER_SIZE = 63;